Integrated the modified example code back into the project
Since almost all features were developed by incrementally modifying
some example code from GitHub, they needed to be merged back into the
main project to not have tons of example items in the project.

The final step in this game jam, which runs until ca. 12:00, is to
finish the implementation of the artillery block by having it placeable
in a custom direction in order to be able to fire into all directions.